requirements.txt
schedule~=0.6.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
requests~=2.23.0
beautifulsoup4~=4.9.3